feat(assets): add loading state to holdings table
