--- packages/blockchain-wallet-v4-frontend/src/scenes/Home/Balances/Table/index.tsx
@@ -1,35 +1,0 @@
-import React from 'react'
-import { connect, ConnectedProps } from 'react-redux'
-import { bindActionCreators } from 'redux'
-
-import { ExtractSuccess } from 'blockchain-wallet-v4/src/types'
-import { actions } from 'data'
-
-import { getData } from './selectors'
-import Template from './template'
-
-class Table extends React.PureComponent<Props> {
-  render() {
-    return this.props.data.cata({
-      Success: val => <Template {...this.props} {...val} />,
-      Failure: () => null,
-      NotAsked: () => null,
-      Loading: () => null
-    })
-  }
-}
-
-const mapStateToProps = state => ({
-  data: getData(state)
-})
-
-const mapDispatchToProps = dispatch => ({
-  actions: bindActionCreators(actions.components.refresh, dispatch)
-})
-
-const connector = connect(mapStateToProps, mapDispatchToProps)
-
-export type SuccessStateType = ExtractSuccess<ReturnType<typeof getData>>
-export type Props = ConnectedProps<typeof connector>
-
-export default connector(Table)
