+++ packages/blockchain-wallet-v4-frontend/src/scenes/Home/Holdings/CoinBalance/template.success.tsx
